requirements.txt
fastmcp
google-adk
python-dotenv
mcp
